requirements.txt
fastapi==0.112.2
uvicorn[standard]==0.30.6
pydantic==2.8.2
python-dotenv==1.0.1
structlog==24.4.0
qdrant-client==1.9.2
pydantic-settings==2.4.0
prometheus-fastapi-instrumentator==6.1.0
openai
numpy==1.26.4
rank-bm25
orjson
pytest
coverage
httpx
ruff
black
